restructuring file hierarchy Co-authored-by: Reese <reesegreen2014@me.com> Co-authored-by: Kim <pekar.kimberly@gmail.com>
